+++ examples/rp235x/async/src/bin/rx_continuous.rs
@@ -1,22 +1,20 @@
-//! This async example demonstrates how to use the LoRa modem in RXCONTINUOUS mode and handle the
-//! RxDone interrupt on DIO0. You will need a second dual_modem chip in range and with the same settings
+//! This example demonstrates how to use the LoRa modem in RXCONTINUOUS mode and handle the RxDone
+//! interrupt on DIO0. You will need a second dual_modem chip in range and with the same settings
 //! to handle tx.
 #![no_std]
 #![no_main]
 
-use core::convert::Infallible;
 use defmt::*;
 use embassy_embedded_hal::shared_bus::asynch::spi::SpiDevice;
-use embassy_embedded_hal::shared_bus::SpiDeviceError;
 use embassy_executor::Spawner;
 use embassy_rp::gpio::{Input, Level, Output, Pull};
 use embassy_rp::peripherals::SPI1;
-use embassy_rp::spi::{Async, Config, Error, Spi};
+use embassy_rp::spi::{Async, Config, Spi};
 use embassy_sync::blocking_mutex::raw::NoopRawMutex;
 use embassy_sync::mutex::Mutex;
 use {defmt_rtt as _, panic_probe as _};
 use sx127xlora::driver::{Sx127xLora, Sx127xLoraConfig};
-use sx127xlora::types::{Dio0Signal, Interrupt};
+use sx127xlora::types::{Dio0Signal, IRQ};
 
 const FREQUENCY_HZ: u32 = 915_000_000;
 
@@ -36,16 +34,16 @@ async fn main(_task_spawner: Spawner) {
 
     let mut config = Sx127xLoraConfig::default();
     config.frequency = FREQUENCY_HZ;
-    let mut sx127x = Sx127xLora::new(spi_dev, config).await.expect("driver init failed :(");
+    let mut sx127x = Sx127xLora::new(spi_dev, config).await.unwrap();
 
-    sx127x.set_dio0(Dio0Signal::RxDone).await.expect("enable_dio0 failed :(");
-    sx127x.receive(None).await.expect("receive failed :(");
+    sx127x.set_dio0(Dio0Signal::RxDone).await.unwrap();
+    sx127x.receive(None).await.unwrap();
 
     loop {
         info!("waiting for RxDone...");
         dio0.wait_for_high().await;
         info!("RxDone triggered!");
-        sx127x.clear_interrupt(Interrupt::RxDone).await.expect("clear interrupt RxDone failed :(");
+        sx127x.clear_irq(IRQ::RxDone).await.unwrap();
         match sx127x.read_rx_data().await {
             Ok(buf) => {
                 let len: usize = buf.iter().filter(|c| **c != 0).count();